fix: preserve send amount selection
