+++ lib/src/features/send/screens/mobile/mobile_send_screen.dart
@@ -16,7 +16,6 @@ import '../../../../core/layout/mobile/mobile_top_nav.dart';
 import '../../../../core/storage/wallet_paths.dart';
 import '../../../../core/theme/app_theme.dart';
 import '../../../../core/widgets/app_button.dart';
-import '../../../../core/widgets/comma_to_dot_input_formatter.dart';
 import '../../../../core/widgets/app_icon.dart';
 import '../../../../core/widgets/app_profile_picture.dart';
 import '../../../../core/widgets/app_toast.dart';
@@ -25,6 +24,7 @@ import '../../../../core/widgets/mobile/mobile_review_row.dart';
 import '../../../../core/widgets/mobile/mobile_surface_card.dart';
 import '../../../../core/widgets/mobile/mobile_tx_fee_info_sheet.dart';
 import '../../../../core/widgets/mobile_text_field.dart';
+import '../../../../core/widgets/sanitizing_decimal_amount_input_formatter.dart';
 import '../../../../providers/account_provider.dart';
 import '../../../../providers/rpc_endpoint_provider.dart';
 import '../../../../providers/sync_provider.dart';
@@ -2290,8 +2290,7 @@ class _MobileSendScreenState extends ConsumerState<MobileSendScreen> {
           _kMobileSendAmountLineHeightPx / _kMobileSendAmountUsdPrefixFontSize,
     );
     final inputFormatters = [
-      const CommaToDotInputFormatter(),
-      _DecimalAmountInputFormatter(
+      SanitizingDecimalAmountInputFormatter(
         maxFractionDigits: _amountInputIsUsd ? 2 : 8,
         maxLength: _amountInputIsUsd ? 12 : 17,
       ),
@@ -3109,54 +3108,6 @@ RenderEditable? _findRenderEditable(RenderObject root) {
     found ??= _findRenderEditable(child);
   });
   return found;
-}
-
-class _DecimalAmountInputFormatter extends TextInputFormatter {
-  const _DecimalAmountInputFormatter({
-    required this.maxFractionDigits,
-    required this.maxLength,
-  });
-
-  final int maxFractionDigits;
-  final int maxLength;
-
-  @override
-  TextEditingValue formatEditUpdate(
-    TextEditingValue oldValue,
-    TextEditingValue newValue,
-  ) {
-    var text = newValue.text;
-    if (text.isEmpty) return newValue;
-
-    final buffer = StringBuffer();
-    var hasDecimal = false;
-    for (final codeUnit in text.codeUnits) {
-      final ch = String.fromCharCode(codeUnit);
-      if (ch == '.') {
-        if (hasDecimal) continue;
-        hasDecimal = true;
-        buffer.write(ch);
-        continue;
-      }
-      if (codeUnit >= 0x30 && codeUnit <= 0x39) {
-        buffer.write(ch);
-      }
-    }
-
-    text = buffer.toString();
-    if (text.startsWith('.')) text = '0$text';
-    if (text.length > maxLength) text = text.substring(0, maxLength);
-    final decimalIndex = text.indexOf('.');
-    if (decimalIndex >= 0) {
-      final maxEnd = decimalIndex + 1 + maxFractionDigits;
-      if (text.length > maxEnd) text = text.substring(0, maxEnd);
-    }
-
-    return TextEditingValue(
-      text: text,
-      selection: TextSelection.collapsed(offset: text.length),
-    );
-  }
 }
 
 class _ReviewZecIcon extends StatelessWidget {
